+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Added `SlevomatCodingStandard.Classes.ClassKeywordOrder` rule to enforce correct order of class modifiers (e.g., `final readonly class`)
+
+### Changed
+- Updated Slevomat Coding Standard to 8.27.*
 
 ## [0.4.2] - 2025-11-28
 ### Changed